Update config.strip for amd64 boot failure in VirtualBox.
This would most likely fatten up the kernel too much to be accepted, but I wanted to put it out for posterity. I spent the time going through line-by-line and group-by-group (ugh) to see which options from radeondrm onward cause a crash during kernel load in VirtualBox 4.3.26.

I also opened a ticket in the hopes that some of the problem is on their end, with a testable disk image: https://www.virtualbox.org/ticket/14035

Of note: None of these changes are needed to boot i386 in the same VirtualBox configuration. Weird.